# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ent.Age:
    properties:
      AGE:
        description: AGE holds the value of the "AGE" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.AgeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AgeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.AgeEdges:
    properties:
      employee:
        description: Employee holds the value of the employee edge.
        items:
          $ref: '#/definitions/ent.Employee'
        type: array
    type: object
  ent.Bookcourse:
    properties:
      ACCESS:
        description: ACCESS holds the value of the "ACCESS" field.
        type: integer
      BOOKTIME:
        description: BOOKTIME holds the value of the "BOOKTIME" field.
        type: string
      DETAIL:
        description: DETAIL holds the value of the "DETAIL" field.
        type: string
      PHONE:
        description: PHONE holds the value of the "PHONE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.BookcourseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BookcourseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.BookcourseEdges:
    properties:
      course:
        $ref: '#/definitions/ent.Course'
        description: Course holds the value of the course edge.
        type: object
      employee:
        $ref: '#/definitions/ent.Employee'
        description: Employee holds the value of the employee edge.
        type: object
      member:
        $ref: '#/definitions/ent.Member'
        description: Member holds the value of the member edge.
        type: object
    type: object
  ent.Classifier:
    properties:
      EQUIPMENTCLASSIFIER:
        description: EQUIPMENTCLASSIFIER holds the value of the "EQUIPMENTCLASSIFIER"
          field.
        type: string
      edges:
        $ref: '#/definitions/ent.ClassifierEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ClassifierQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ClassifierEdges:
    properties:
      equipment:
        description: Equipment holds the value of the equipment edge.
        items:
          $ref: '#/definitions/ent.Equipment'
        type: array
    type: object
  ent.Course:
    properties:
      COURSE:
        description: COURSE holds the value of the "COURSE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CourseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CourseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CourseEdges:
    properties:
      bookcourse:
        description: Bookcourse holds the value of the bookcourse edge.
        items:
          $ref: '#/definitions/ent.Bookcourse'
        type: array
    type: object
  ent.Employee:
    properties:
      EMPLOYEEADDRESS:
        description: EMPLOYEEADDRESS holds the value of the "EMPLOYEEADDRESS" field.
        type: string
      EMPLOYEEID:
        description: EMPLOYEEID holds the value of the "EMPLOYEEID" field.
        type: string
      EMPLOYEENAME:
        description: EMPLOYEENAME holds the value of the "EMPLOYEENAME" field.
        type: string
      IDCARDNUMBER:
        description: IDCARDNUMBER holds the value of the "IDCARDNUMBER" field.
        type: string
      edges:
        $ref: '#/definitions/ent.EmployeeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EmployeeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.EmployeeEdges:
    properties:
      age:
        $ref: '#/definitions/ent.Age'
        description: Age holds the value of the age edge.
        type: object
      bookcourse:
        description: Bookcourse holds the value of the bookcourse edge.
        items:
          $ref: '#/definitions/ent.Bookcourse'
        type: array
      equipment:
        description: Equipment holds the value of the equipment edge.
        items:
          $ref: '#/definitions/ent.Equipment'
        type: array
      equipmentrental:
        description: Equipmentrental holds the value of the equipmentrental edge.
        items:
          $ref: '#/definitions/ent.Equipmentrental'
        type: array
      payment:
        description: Payment holds the value of the payment edge.
        items:
          $ref: '#/definitions/ent.Payment'
        type: array
      position:
        $ref: '#/definitions/ent.Position'
        description: Position holds the value of the position edge.
        type: object
      promotion:
        description: Promotion holds the value of the promotion edge.
        items:
          $ref: '#/definitions/ent.Promotion'
        type: array
      salary:
        $ref: '#/definitions/ent.Salary'
        description: Salary holds the value of the salary edge.
        type: object
    type: object
  ent.Equipment:
    properties:
      EQUIPMENTAMOUNT:
        description: EQUIPMENTAMOUNT holds the value of the "EQUIPMENTAMOUNT" field.
        type: integer
      EQUIPMENTDATE:
        description: EQUIPMENTDATE holds the value of the "EQUIPMENTDATE" field.
        type: string
      EQUIPMENTDETAIL:
        description: EQUIPMENTDETAIL holds the value of the "EQUIPMENTDETAIL" field.
        type: string
      EQUIPMENTNAME:
        description: EQUIPMENTNAME holds the value of the "EQUIPMENTNAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.EquipmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EquipmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.EquipmentEdges:
    properties:
      classifier:
        $ref: '#/definitions/ent.Classifier'
        description: Classifier holds the value of the classifier edge.
        type: object
      employee:
        $ref: '#/definitions/ent.Employee'
        description: Employee holds the value of the employee edge.
        type: object
      equipmentrental:
        description: Equipmentrental holds the value of the equipmentrental edge.
        items:
          $ref: '#/definitions/ent.Equipmentrental'
        type: array
      equipmenttype:
        $ref: '#/definitions/ent.Equipmenttype'
        description: Equipmenttype holds the value of the equipmenttype edge.
        type: object
      zone:
        $ref: '#/definitions/ent.Zone'
        description: Zone holds the value of the zone edge.
        type: object
    type: object
  ent.Equipmentrental:
    properties:
      RENTALAMOUNT:
        description: RENTALAMOUNT holds the value of the "RENTALAMOUNT" field.
        type: integer
      RENTALDATE:
        description: RENTALDATE holds the value of the "RENTALDATE" field.
        type: string
      RETURNDATE:
        description: RETURNDATE holds the value of the "RETURNDATE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.EquipmentrentalEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EquipmentrentalQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.EquipmentrentalEdges:
    properties:
      employee:
        $ref: '#/definitions/ent.Employee'
        description: Employee holds the value of the employee edge.
        type: object
      equipment:
        $ref: '#/definitions/ent.Equipment'
        description: Equipment holds the value of the equipment edge.
        type: object
      equipmenttype:
        $ref: '#/definitions/ent.Equipmenttype'
        description: Equipmenttype holds the value of the equipmenttype edge.
        type: object
      member:
        $ref: '#/definitions/ent.Member'
        description: Member holds the value of the member edge.
        type: object
    type: object
  ent.Equipmenttype:
    properties:
      EQUIPMENTTYPE:
        description: EQUIPMENTTYPE holds the value of the "EQUIPMENTTYPE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.EquipmenttypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EquipmenttypeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.EquipmenttypeEdges:
    properties:
      equipment:
        description: Equipment holds the value of the equipment edge.
        items:
          $ref: '#/definitions/ent.Equipment'
        type: array
      equipmentrental:
        description: Equipmentrental holds the value of the equipmentrental edge.
        items:
          $ref: '#/definitions/ent.Equipmentrental'
        type: array
    type: object
  ent.Member:
    properties:
      MEMBERID:
        description: MEMBERID holds the value of the "MEMBERID" field.
        type: string
      MEMBERNAME:
        description: MEMBERNAME holds the value of the "MEMBERNAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.MemberEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MemberQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.MemberEdges:
    properties:
      bookcourse:
        description: Bookcourse holds the value of the bookcourse edge.
        items:
          $ref: '#/definitions/ent.Bookcourse'
        type: array
      equipmentrental:
        description: Equipmentrental holds the value of the equipmentrental edge.
        items:
          $ref: '#/definitions/ent.Equipmentrental'
        type: array
      payment:
        description: Payment holds the value of the payment edge.
        items:
          $ref: '#/definitions/ent.Payment'
        type: array
    type: object
  ent.Payment:
    properties:
      EMAIL:
        description: EMAIL holds the value of the "EMAIL" field.
        type: string
      PAYMENTAMOUNT:
        description: PAYMENTAMOUNT holds the value of the "PAYMENTAMOUNT" field.
        type: string
      PAYMENTDATE:
        description: PAYMENTDATE holds the value of the "PAYMENTDATE" field.
        type: string
      PHONENUMBER:
        description: PHONENUMBER holds the value of the "PHONENUMBER" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PaymentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PaymentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PaymentEdges:
    properties:
      employee:
        $ref: '#/definitions/ent.Employee'
        description: Employee holds the value of the employee edge.
        type: object
      member:
        $ref: '#/definitions/ent.Member'
        description: Member holds the value of the member edge.
        type: object
      paymenttype:
        $ref: '#/definitions/ent.Paymenttype'
        description: Paymenttype holds the value of the paymenttype edge.
        type: object
      promotion:
        $ref: '#/definitions/ent.Promotion'
        description: Promotion holds the value of the promotion edge.
        type: object
    type: object
  ent.Paymenttype:
    properties:
      TYPE:
        description: TYPE holds the value of the "TYPE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PaymenttypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PaymenttypeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PaymenttypeEdges:
    properties:
      payment:
        description: Payment holds the value of the payment edge.
        items:
          $ref: '#/definitions/ent.Payment'
        type: array
    type: object
  ent.Position:
    properties:
      POSITION:
        description: POSITION holds the value of the "POSITION" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PositionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PositionQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PositionEdges:
    properties:
      employee:
        description: Employee holds the value of the employee edge.
        items:
          $ref: '#/definitions/ent.Employee'
        type: array
    type: object
  ent.Promotion:
    properties:
      CODE:
        description: CODE holds the value of the "CODE" field.
        type: string
      DATE:
        description: DATE holds the value of the "DATE" field.
        type: string
      DESC:
        description: DESC holds the value of the "DESC" field.
        type: string
      NAME:
        description: NAME holds the value of the "NAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PromotionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PromotionQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PromotionEdges:
    properties:
      employee:
        $ref: '#/definitions/ent.Employee'
        description: Employee holds the value of the employee edge.
        type: object
      payment:
        description: Payment holds the value of the payment edge.
        items:
          $ref: '#/definitions/ent.Payment'
        type: array
      promotionamount:
        $ref: '#/definitions/ent.Promotionamount'
        description: Promotionamount holds the value of the promotionamount edge.
        type: object
      promotiontype:
        $ref: '#/definitions/ent.Promotiontype'
        description: Promotiontype holds the value of the promotiontype edge.
        type: object
    type: object
  ent.Promotionamount:
    properties:
      AMOUNT:
        description: AMOUNT holds the value of the "AMOUNT" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.PromotionamountEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PromotionamountQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PromotionamountEdges:
    properties:
      promotion:
        description: Promotion holds the value of the promotion edge.
        items:
          $ref: '#/definitions/ent.Promotion'
        type: array
    type: object
  ent.Promotiontype:
    properties:
      TYPE:
        description: TYPE holds the value of the "TYPE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PromotiontypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PromotiontypeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PromotiontypeEdges:
    properties:
      promotion:
        description: Promotion holds the value of the promotion edge.
        items:
          $ref: '#/definitions/ent.Promotion'
        type: array
    type: object
  ent.Salary:
    properties:
      SALARY:
        description: SALARY holds the value of the "SALARY" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.SalaryEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SalaryQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.SalaryEdges:
    properties:
      employee:
        description: Employee holds the value of the employee edge.
        items:
          $ref: '#/definitions/ent.Employee'
        type: array
    type: object
  ent.Zone:
    properties:
      EQUIPMENTZONE:
        description: EQUIPMENTZONE holds the value of the "EQUIPMENTZONE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ZoneEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ZoneQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ZoneEdges:
    properties:
      equipment:
        description: Equipment holds the value of the equipment edge.
        items:
          $ref: '#/definitions/ent.Equipment'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API Patient
  version: "1.0"
paths:
...
  /classifiers:
    get:
      description: list classifier entities
      operationId: list-classifier
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Classifier'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List classifier entities
    post:
      consumes:
      - application/json
      description: Create classifier
      operationId: create-classifier
      parameters:
      - description: Classifier entity
        in: body
        name: classifier
        required: true
        schema:
          $ref: '#/definitions/ent.Classifier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Classifier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create classifier
  /classifiers/{id}:
    get:
      description: get classifier by ID
      operationId: get-classifier
      parameters:
      - description: Classifier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Classifier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a classifier entity by ID
...
  /equipmentbyemployees/{id}:
    get:
      description: get GetEquipmentbyEmployee by ID
      operationId: get-GetEquipmentbyEmployee
      parameters:
      - description: GetEquipmentbyEmployee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Equipment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a GetEquipmentbyEmployee entity by ID
...
  /equipments:
    get:
      description: list equipment entities
      operationId: list-equipment
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Equipment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List equipment entities
    post:
      consumes:
      - application/json
      description: Create equipment
      operationId: create-equipment
      parameters:
      - description: Equipment entity
        in: body
        name: equipment
        required: true
        schema:
          $ref: '#/definitions/ent.Equipment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Equipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create equipment
  /equipments/{id}:
    get:
      description: get equipment by ID
      operationId: get-equipment
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Equipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a equipment entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"